удален неиспользуемый код с прошлой домашки

+++ app/src/main/java/com/example/playlist_maker_dev/search/data/repository/TracksRepositoryImpl.kt
@@ -26,7 +26,7 @@ class TracksRepositoryImpl(private val networkClient: NetworkClient) : TracksRep
                         trackName = trackDto.trackName,
                         artistName = trackDto.artistName,
                         trackTimeMillis = SimpleDateFormat("mm:ss", Locale.getDefault()).format(
-                            trackDto.trackTimeMillis.toLong()
+                            trackDto.trackTimeMillis
                         ),
                         artworkUrl100 = trackDto.artworkUrl100,
                         collectionName = trackDto.collectionName,
